# Voting Key generation using typescript and multi file support