--- a/cmds/config-testnet-supernode.sh
+++ b/cmds/config-testnet-supernode.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+symbol-bootstrap config -p testnet -a dual -t target/testnet-supernode -c test/supernode.yml --noPassword $1 $2 $3