--- a/cmds/link-testnet-supernode-external-node.sh
+++ b/cmds/link-testnet-supernode-external-node.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-symbol-bootstrap link -t target/testnet-supernode --useKnownRestGateways $1 $2 $3 $4 $5 $6 $7
+symbol-bootstrap link -t target/testnet-supernode --useKnownRestGateways --password 1111 $1 $2 $3 $4 $5 $6 $7