--- a/cmds/stop-testnet-supernode.sh
+++ b/cmds/stop-testnet-supernode.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+# docker rm -f $(docker ps -aq)
+symbol-bootstrap stop -t target/testnet-supernode