--- a/cmds/upgradeVotingKeys-testnet-supernode.sh
+++ b/cmds/upgradeVotingKeys-testnet-supernode.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+symbol-bootstrap upgradeVotingKeys -t target/testnet-supernode $1 $2 $3