Update:

This PR now supports both the original catapult tools' voting key file generation and the native typescript. The original is the one used by default. You can enable the new native typescript with the useExperimentalNativeVotingKeyGeneration: true custom preset.

I'm also working on multi-voting file support. The voting file metadata (public key, start, and end epoch) is loaded directly from the .dat files. 

Voting file upgrade idea:

1. Bootstrap knows the network "current" finalizationEpoch (by either a manually upgraded custom preset value for offline setup or asking the network for online)
2. If the node's current voting file has an end epoch close to the current epoch (close to expiring), --upgrade will create 'private_key_tree2.dat' that continues the current file. I defined "Close to expiring"  when the epoch is in the upper half of the voting file. If the file's epoch length is 360, close to expiring will be 180+. Our nodes are 180 epochs, current epoch is 104 so they are close to expiring.
3. When a new voting file is created, bootstrap will advise running the link command again. The link command detects what has changed and announces the required transactions (linking the new voting key). The voting key files metadata (public key, start, and end epochs) will be in addresses.yml so it can also be linked using the symbol CLI if desired. 

For the end-user, the upgrade process is pretty much the same:

```
$ npm install -g symbol-bootstrap
$ symbol-bootstrap start -p mainnet -a dual -c custom-preset.yml --upgrde
>>> New Voting File Created!
$ symbol-bootstrap link
```

We can also add a new command "upgradeVotingFiles" that does the --upgrade of voting files explicitly 

TODOs:
- Test how the node behaves when transitioning from one file to another
- Double-check if we need to unlink already expired voting keys in the network. I think there is a limit about how many public keys can be linked. We may only need the current voting key and the future voting key
- Double-check want to do with old voting key files. Do we need to remove them? If we need to remove them, we need to "shift them", if voting file 1 is removed, voting file 2 should be renamed to voting key 1, 3 to voting key 2, etc

Slack chat: https://nemgroup-workspace.slack.com/archives/C015AEQ9ZL2/p1616410897191200

Related to https://github.com/nemtech/symbol-bootstrap/issues/149

Original:
This is my first implementation of voting key file generation done in Javascript/Typescript. 

I think the code is ok, the main issue is that the SDK is really slow generating keys and signing data. The SDK uses [tweetnacl](https://www.npmjs.com/package/tweetnacl), which is small but slow. These are some numbers on my dev machine when creating the voting key file:

(votingKeyEndEpoch - votingKeyStartEpoch + 1 -> TIME)
- 1 epoch -> 50 MS
- 5 epochs -> 114 MS
- 1000 epochs -> 10 seconds
- 10000 epochs -> 1 minute 41 seconds 
- 26280 epochs -> 4 Minutes 25 seconds

@gimer , could you validate the logic in https://github.com/nemtech/symbol-bootstrap/blob/native_voting_key_file/src/service/VotingUtils.ts#L24?

As unit testing, I'm comparing files generates with catapult tools against files generated using the native implementation. Checking that the headers and sizes are the same. I would add more assertions validating the internal items (private key and signatures). 

To speed up the implementation we could:

- Change the SDK crypto library. It could make the SDK faster but it would make it heavier, it may not work on Web, it could have secondary effects. Not Recommended at this stage.
- Change the crypto lib Bootstrap uses for the voting key implementation but keep the SDK for the rest of the key generation. Maybe something like [monocypher](https://monocypher.org/speed)

Before doing any improvement, I want to double-check the logic is fine. 

I'm pushing hard for this feature not because I want to reimplement the tool. This will allow me to make Bootstrap config offline without docker and OS independent (catapult tools wouldn't need to be compiled and installed in the offline machine). Supernodes will have fewer excuses to not use Bootstrap to setup the configuration (when runing barebone or with docker compose in the final online node machine)

Fixes https://github.com/nemtech/symbol-bootstrap/issues/169
Related to https://github.com/nemtech/symbol-bootstrap/issues/170

Update:
noble-ed25519 takes ~40 seconds for the 26280 epochs