@fboucquez 2021-03-02
> 10000 - I hope you remember that for publicnet limit is 1460, so the tool should probably check that limit somewhere
Thanks. We also use bootstrap for private networks,  epoch length would be large in this case. I can add the limit the public/testnet nodes