@fboucquez 2021-03-02
Update, I've added multi implementation (all implementations pass the vector tests). The fastest one by far is noble-ed25519